requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
altair>=5.0.0
streamlit-autorefresh>=1.0.1
plotly>=5.0.0
plotly>=5.17.0
python-dateutil>=2.8.2
pytz>=2023.3
pyyaml>=6.0
python-json-logger>=2.0.7
python-dotenv>=1.0.0
cryptography>=41.0.0
cachetools>=5.3.0
joblib>=1.3.0
streamlit>=1.28.0
pandas>=2.0.0
finnhub-python>=2.4.0
yfinance>=0.2.28
requests>=2.31.0
altair>=5.0.0
streamlit-autorefresh>=0.0.1
streamlit>=1.28.0
pandas>=2.0.0
finnhub-python>=2.4.0
yfinance>=0.2.28
requests>=2.31.0
altair>=5.0.0
streamlit-autorefresh>=0.0.1
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.21.0
